+++ src/icemac/ab/calendar/browser/calendar.py
@@ -28,6 +28,7 @@ class Dispatcher(icemac.ab.calendar.browser.base.View):
     """Dispatch to month resp. year view."""
 
     possible_views = {'month': 'month.html',
+                      'month-list': 'month-list.html',
                       'year': 'year.html'}
 
     def __call__(self):
@@ -215,6 +216,12 @@ class MonthCalendar(TabularCalendar):
         return u'<h2 class="no-screen">{month} {year}</h2>\n{cal}'.format(
             month=self.get_month_name(self.month), year=self.month.year,
             cal=self.renderer())
+
+
+class MonthListCalendar(MonthCalendar):
+    """List display of a month calendar. (like the print style)"""
+
+    css_class = 'month-list'
 
 
 class IYearSelector(zope.interface.Interface):